# Conflicts:
#	.github/workflows/linux-build.yml
#	src/core/DynaRec_aa64/recompiler.cc
#	src/core/DynaRec_x64/recompiler.cc
#	src/core/cdrom.cc
#	src/core/r3000a.cc
#	src/core/r3000a.h
#	src/core/sio.h
#	src/core/sstate.h
#	src/gui/widgets/registers.cc

+++ src/core/sio.h
@@ -233,7 +233,7 @@ class SIO {
 
     bool isReceiveIRQReady();
     bool isTransmitReady();
-    static inline void scheduleInterrupt(uint32_t eCycle) {
+    static inline void scheduleInterrupt(uint64_t eCycle) {
         g_emulator->m_cpu->schedule(Schedule::SIO, eCycle);
 #if 0
 // Breaks Twisted Metal 2 intro
